08072019 AN Update
The selection program was modified to take into account only records for which the same station has Mainshock and Aftershock. Also now the files can be accessed to the Kowalsky Group Folder and the PEER Website further developments would allow to do both at the same time. The plan is to share with the rest of the Group.